requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
langchain
langchain-community
langchain-openai
chromadb
pypdf
tiktoken